+++ PierresOrderTracker/Models/Order.cs
@@ -9,11 +9,11 @@ namespace PierresOrderTracker.Models
     public string Title { get; set; }
     public int Id { get; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
     public string Date { get; set; }
     private static List<Order> _instances = new List<Order> { };
 
-    public Order(string title, string description, int price, string date)
+    public Order(string title, string description, decimal price, string date)
     {
       Title = title;
       Description = description;
